+++ app/controllers/ckeditor/application_controller.rb
@@ -9,9 +9,8 @@ class Ckeditor::ApplicationController < Ckeditor.parent_controller.constantize
 
   def respond_with_asset(asset)
     asset_response = Ckeditor::AssetResponse.new(asset, request)
-    callback = ckeditor_before_create_asset(asset)
 
-    if callback && asset.save
+    if asset.save
       render asset_response.success(config.relative_url_root)
     else
       render asset_response.errors