+++ lib/ckeditor/orm/active_record.rb
@@ -13,8 +13,6 @@ module Ckeditor
           def self.extended(base)
             base.class_eval do
               self.table_name = 'ckeditor_assets'
-
-              belongs_to :assetable, polymorphic: true
             end
           end
         end