+++ lib/generators/ckeditor/templates/active_record/carrierwave/migration.rb
@@ -4,9 +4,6 @@ class CreateCkeditorAssets < ActiveRecord::Migration
       t.string  :data_file_name, null: false
       t.string  :data_content_type
       t.integer :data_file_size
-
-      t.integer :assetable_id
-      t.string  :assetable_type, limit: 30
       t.string  :type, limit: 30
 
       # Uncomment it to save images dimensions, if your need it
@@ -16,8 +13,7 @@ class CreateCkeditorAssets < ActiveRecord::Migration
       t.timestamps null: false
     end
 
-    add_index :ckeditor_assets, [:assetable_type, :type, :assetable_id], name: :idx_ckeditor_assetable_type
-    add_index :ckeditor_assets, [:assetable_type, :assetable_id], name: :idx_ckeditor_assetable
+    add_index :ckeditor_assets, :type
   end
 
   def self.down