+++ lib/generators/ckeditor/templates/active_record/dragonfly/migration.rb
@@ -5,9 +5,6 @@ class CreateCkeditorAssets < ActiveRecord::Migration
       t.string  :data_name, null: false
       t.string  :data_mime_type
       t.integer :data_size
-
-      t.integer :assetable_id
-      t.string  :assetable_type, limit: 30
       t.string  :type, limit: 30
 
       # Uncomment these to save image dimensions, if your need them.
@@ -17,8 +14,7 @@ class CreateCkeditorAssets < ActiveRecord::Migration
       t.timestamps null: false
     end
 
-    add_index :ckeditor_assets, [:assetable_type, :type, :assetable_id], name: :idx_ckeditor_assetable_type
-    add_index :ckeditor_assets, [:assetable_type, :assetable_id], name: :idx_ckeditor_assetable
+    add_index :ckeditor_assets, :type
   end
 
   def self.down