+++ lib/generators/ckeditor/templates/active_record/refile/migration.rb
@@ -5,9 +5,6 @@ class CreateCkeditorAssets < ActiveRecord::Migration
       t.string  :data_filename, null: false
       t.integer :data_size
       t.string  :data_content_type
-
-      t.integer :assetable_id
-      t.string  :assetable_type, limit: 30
       t.string  :type, limit: 30
 
       # Uncomment it to save images dimensions, if your need it
@@ -17,8 +14,7 @@ class CreateCkeditorAssets < ActiveRecord::Migration
       t.timestamps null: false
     end
 
-    add_index :ckeditor_assets, [:assetable_type, :type, :assetable_id], name: :idx_ckeditor_assetable_type
-    add_index :ckeditor_assets, [:assetable_type, :assetable_id], name: :idx_ckeditor_assetable
+    add_index :ckeditor_assets, :type
   end
 
   def self.down